+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.2
 * Add clean commands for all languages
+* Handle language containing dash
 
 ### 0.0.1
 * Initial version